feat: Add resource group name variable to Terraform configuration
This change adds a new `resource_group_name` variable to the Terraform configuration for the XWiki deployment. This variable allows the resource group name to be specified as an input, providing more flexibility and control over the deployment. The `main.tf` file has been updated to use the new variable for the resource group name, instead of generating it from a local prefix.