feat: Migrate to use existing resource group
This change updates the Terraform configuration to use an existing resource group instead of creating a new one. The `azurerm_resource_group` resource has been replaced with a `data` source to reference the existing resource group. This ensures the infrastructure is deployed to the correct resource group without creating a new one.

The changes include:

- Replacing the `azurerm_resource_group` resource with a `data` source for `azurerm_resource_group`
- Updating the references to the resource group in other resources to use the data source
- Removing the unused `location` and `tags` arguments from the `azurerm_resource_group` data source

BREAKING CHANGE: This change requires updating the input variables to use the new `resource_group_name` variable instead of the previous `location` and `resource_group_name` variables.